requirements.txt
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
imageio==2.33.1
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
python-engineio==4.8.0
requests==2.31.0
scipy==1.11.4
simple-websocket==1.0.0
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0